CB-17716 Introduce the brand new DTO and views classes, Views added to the related entity as interface. Dedicated MDC info provider interface introduced

+++ core/src/main/java/com/sequenceiq/cloudbreak/converter/v4/clustertemplate/ClusterTemplateToClusterTemplateV4ResponseConverter.java
@@ -52,7 +52,7 @@ public class ClusterTemplateToClusterTemplateV4ResponseConverter {
                     stackV4Request = null;
                 }
                 clusterTemplateV4Response.setDistroXTemplate(getIfNotNull(stackV4Request, stackV4RequestConverter::convert));
-                clusterTemplateV4Response.setNodeCount(stack.get().getFullNodeCount());
+                clusterTemplateV4Response.setNodeCount(stack.get().getFullNodeCount().intValue());
             }
         } else {
             try {
